Refactor repository tests to ensure a common interface
