+++ graphql-api/src/persistence/fake-db.js
@@ -1,13 +1,20 @@
 const { Haiku } = require('../domain/types/Haiku');
 
 class FakeHaikuDB {
-
   async init() {
     this.haikuMap = {};
     return new Promise((resolve) => {
       resolve();
     });
   }
+
+  /* eslint-disable class-methods-use-this */
+  async close() {
+    return new Promise((resolve) => {
+      resolve();
+    });
+  }
+  /* eslint-enable class-methods-use-this */
 
   createHaiku(haikuInput) {
     const { serverId } = haikuInput;
@@ -28,10 +35,12 @@ class FakeHaikuDB {
   }
 
   getHaiku(serverId, id) {
-    if (!this.haikuMap[serverId] || !this.haikuMap[serverId][id]) {
-      throw new Error(`No haiku with id ${id} found in server ${serverId}`);
-    }
-    return new Haiku(id, this.haikuMap[serverId][id]);
+    return new Promise((resolve, reject) => {
+      if (!this.haikuMap[serverId] || !this.haikuMap[serverId][id]) {
+        reject(new Error(`No haiku with id ${id} found in server ${serverId}`));
+      }
+      resolve(new Haiku(id, this.haikuMap[serverId][id]));
+    });
   }
 
   clearHaiku(serverId, id) {
@@ -44,6 +53,7 @@ class FakeHaikuDB {
     this.haikuMap = {};
   }
 
+  /* eslint-disable class-methods-use-this */
   getChannel(id) {
     return { id };
   }
@@ -51,6 +61,7 @@ class FakeHaikuDB {
   getServer(id) {
     return { id };
   }
+  /* eslint-enable class-methods-use-this */
 }
 
 exports.FakeHaikuDB = FakeHaikuDB;
