+++ graphql-api/test/persistence/sqlite.js
@@ -1,0 +1,5 @@
+const { SQLiteHaikuDB } = require('../../src/persistence/sqlite');
+const { testRepo } = require('./repo-tests');
+
+const repo = new SQLiteHaikuDB();
+testRepo(repo, 'SQLite DB');
